+++ Changelog.md
@@ -9,6 +9,12 @@ Added `"rust_1_55"` and `"rust_1_56"` features.
 Change: functions that required the `"deref_raw_in_fn"` feature can be used on Rust 1.56.0 (whenever that reaches stable) with the `"rust_1_56"` feature.
 
 Added these slice functions
+- `konst::slice::bytes_trim`
+- `konst::slice::bytes_trim_start`
+- `konst::slice::bytes_trim_end`
+- `konst::slice::bytes_trim_matches`
+- `konst::slice::bytes_trim_start_matches`
+- `konst::slice::bytes_trim_end_matches`
 - `konst::slice::get_from_mut` (requires the `"mut_refs"` feature)
 - `konst::slice::get_from`
 - `konst::slice::get_mut` (requires the `"mut_refs"` feature)
@@ -17,6 +23,7 @@ Added these slice functions
 - `konst::slice::get_up_to_mut` (requires the `"mut_refs"` feature)
 - `konst::slice::get_up_to`
 - `konst::slice::get`
+
 
 Added these str functions, all of which require the `"rust_1_55"` feature:
 - `konst::string::from_utf8`
@@ -29,6 +36,12 @@ Added these str functions, all of which require the `"rust_1_55"` feature:
 - `konst::string::str_strip_prefix`
 - `konst::string::str_strip_suffix`
 - `konst::string::str_up_to`
+- `konst::string::trim`
+- `konst::string::trim_start`
+- `konst::string::trim_end`
+- `konst::string::trim_matches`
+- `konst::string::trim_start_matches`
+- `konst::string::trim_end_matches`
 
 Added the `konst::string::from_utf8` macro.